+++ src/main/java/xyz/suchdoge/webapi/model/blockchain/Balance.java
@@ -4,18 +4,16 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
 
-import java.math.BigDecimal;
-
 @Getter
 public class Balance {
-    private final BigDecimal availableBalance;
-    private final BigDecimal pendingReceivedBalance;
+    private final Double availableBalance;
+    private final Double pendingReceivedBalance;
     private final Network network;
 
     @JsonCreator
     public Balance(
-            @JsonProperty("available_balance") BigDecimal availableBalance,
-            @JsonProperty("pending_received_balance") BigDecimal pendingReceivedBalance,
+            @JsonProperty("available_balance") Double availableBalance,
+            @JsonProperty("pending_received_balance") Double pendingReceivedBalance,
             @JsonProperty("network") String network
     ) {
         this.availableBalance = availableBalance;
