+++ src/main/java/xyz/suchdoge/webapi/model/blockchain/Wallet.java
@@ -3,21 +3,19 @@ package xyz.suchdoge.webapi.model.blockchain;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
 
-import java.math.BigDecimal;
-
 @Getter
 public class Wallet {
     private final String address;
     private final String label;
-    private final BigDecimal availableBalance;
-    private final BigDecimal pendingReceivedBalance;
+    private final Double availableBalance;
+    private final Double pendingReceivedBalance;
     private final Long userId;
     private final Network network;
 
     public Wallet(@JsonProperty("address") String address,
                   @JsonProperty("label") String label,
-                  @JsonProperty("available_balance") BigDecimal availableBalance,
-                  @JsonProperty("pending_received_balance") BigDecimal pendingReceivedBalance,
+                  @JsonProperty("available_balance") Double availableBalance,
+                  @JsonProperty("pending_received_balance") Double pendingReceivedBalance,
                   @JsonProperty("user_id") Long userId,
                   @JsonProperty("network") String network) {
         this.address = address;
